pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "delphi"
version = "0.1.0"
description = "Delphi research code"
requires-python = ">=3.12"

# Optional, but good hygiene:
readme = "README.md"
authors = [{ name = "Andrew Williams" }]
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.15",
    "aiosignal==1.4.0",
    "annotated-types==0.7.0",
    "anthropic==0.39.0",
    "anyio==4.9.0",
    "asttokens==3.0.0",
    "attrs==25.3.0",
    "certifi==2025.7.14",
    "comm==0.2.2",
    "contourpy==1.3.2",
    "cycler==0.12.1",
    "debugpy==1.8.15",
    "decorator==5.2.1",
    "distro==1.9.0",
    "executing==2.2.0",
    "filelock==3.19.1",
    "fonttools==4.59.0",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "fuzzywuzzy==0.18.0",
    "groq==0.31.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ipykernel==6.30.0",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "joblib==1.5.1",
    "jsonlines==4.0.0",
    "jupyter-client==8.6.3",
    "jupyter-core==5.8.1",
    "kiwisolver==1.4.8",
    "levenshtein==0.27.1",
    "lightning-utilities==0.15.2",
    "loguru==0.7.3",
    "markupsafe==3.0.2",
    "matplotlib==3.10.3",
    "matplotlib-inline==0.1.7",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "nest-asyncio==1.6.0",
    "networkx==3.5",
    "numpy==2.3.1",
    "nvidia-cublas-cu12==12.8.4.1",
    "nvidia-cuda-cupti-cu12==12.8.90",
    "nvidia-cuda-nvrtc-cu12==12.8.93",
    "nvidia-cuda-runtime-cu12==12.8.90",
    "nvidia-cudnn-cu12==9.10.2.21",
    "nvidia-cufft-cu12==11.3.3.83",
    "nvidia-cufile-cu12==1.13.1.3",
    "nvidia-curand-cu12==10.3.9.90",
    "nvidia-cusolver-cu12==11.7.3.90",
    "nvidia-cusparse-cu12==12.5.8.93",
    "nvidia-cusparselt-cu12==0.7.1",
    "nvidia-nccl-cu12==2.27.3",
    "nvidia-nvjitlink-cu12==12.8.93",
    "nvidia-nvtx-cu12==12.8.90",
    "openai==1.98.0",
    "packaging==25.0",
    "pandas==2.3.1",
    "parso==0.8.4",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "platformdirs==4.3.8",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pygments==2.19.2",
    "pyparsing==3.2.3",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.0",
    "python-levenshtein==0.27.1",
    "pytorch-lightning==2.5.3",
    "pytz==2025.2",
    "pyyaml==6.0.1",
    "pyzmq==27.0.0",
    "rapidfuzz==3.13.0",
    "ruff==0.12.8",
    "scikit-learn==1.7.1",
    "scipy==1.16.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "sympy==1.14.0",
    "threadpoolctl==3.6.0",
    "torch==2.8.0",
    "torchmetrics==1.8.1",
    "tornado==6.5.1",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "triton==3.4.0",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "wcwidth==0.2.13",
    "yarl==1.20.1",
    "flask==3.0.3",
]

[tool.setuptools]
# default package_dir is ".", which matches your flat layout
py-modules = ["expert", "icl_delphi_tests", "icl_initial_forecasts", "mediator"]

[tool.setuptools.packages.find]
# ✅ list ONLY real Python packages you want importable
# (adjust if you have more code dirs)
include = [
  "agents*",
  "dataset*",
  "genetic_evolution*",
  "analyze*",
  "prompt_learning*",
  "runners*",
  "scripts*",
  "utils*"
]

# 🚫 exclude outputs, configs, archives, plots, etc.
exclude = [
  "archive*",
  "configs*",
  "results*",
  "results_extra*",
  "plots_output*",
  "outputs*",
  "outputs_bak*",
  "initial_*",
  "prompt_learning/outputs*",
]
